Core.html
This file is the core of the framework. The viewer will run from this
file inside the app. I have prepared integrated CSS and JavaScript for
use later. JQuery has been linked up and I have included a function
which allows for the extraction of variables from a URL (will be used
for navigation purposes within the app).